Split pool module in three submodules
